requirements.txt
aiogram==2.11.2
aiohttp==3.7.3
aioredis==1.3.1
async-timeout==3.0.1
asyncpg==0.21.0
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
chardet==3.0.4
gino==1.0.1
hiredis==1.1.0
idna==2.10
multidict==5.1.0
python-dotenv==0.15.0
pytz==2020.4
SQLAlchemy==1.3.22
typing-extensions==3.7.4.3
yarl==1.6.3
